memory: mediatek: Fix the error on casting void pointer to int
When using clang with options '-Werror,-Wvoid-pointer-to-int-cast',
there is the compile error.
	error: cast to smaller integer type 'int' from 'const void *'

Use the C99 uintptr_t to cast first.

(cherry-picked from commit f5311129753dbb52a2b9e8e975293a30217451c9)
CR-Id: ALPS05539426
Feature: External Memory Interface(emi)

+++ drivers/memory/mediatek/emi-cen.c
@@ -1121,7 +1121,7 @@ static int emicen_probe(struct platform_device *pdev)
 	if (!cen)
 		return -ENOMEM;
 
-	cen->ver = (int)of_device_get_match_data(&pdev->dev);
+	cen->ver = (int)(uintptr_t)of_device_get_match_data(&pdev->dev);
 
 	ret = of_property_read_u32(emicen_node,
 		"ch_cnt", &(cen->ch_cnt));
